+++ src/main/java/persistence/entity/SimpleEntityManager.java
@@ -6,7 +6,7 @@ import persistence.sql.dialect.Dialect;
 import persistence.sql.dml.DeleteQueryBuilder;
 import persistence.sql.dml.InsertQueryBuilder;
 import persistence.sql.dml.SelectQueryBuilder;
-import persistence.sql.dml.WhereRecord;
+import persistence.sql.dml.conditions.WhereRecord;
 
 import java.util.List;
