+++ src/main/java/persistence/sql/dml/conditions/WhereCondition.java
@@ -1,4 +1,4 @@
-package persistence.sql.dml;
+package persistence.sql.dml.conditions;
 
 import persistence.sql.metadata.ColumnMetadata;
